Refactoring to handle clean installation (without exporting to PYTHONPATH)
