+++ avatar/game_avatar_custom.py
@@ -3,7 +3,7 @@ from models.captioning.evaluate import get_eval_captioning_model
 # from models.vqa.evaluate_vqa import get_eval_vqa_model
 from models.vqa.evaluate_attention_vqa import get_eval_vqa_model
 import tensorflow as tf
-from models.utils.util import get_config
+from config.util import get_config
